Backend/OfficeSpaces/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==2.2.10
django-cors-headers==3.2.1
djangorestframework==3.11.0
pathspec==0.7.0
pytz==2019.3
regex==2020.1.8
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
pillow==6.2.2
requests==2.23.0
pandas==1.0.1
pandas-datareader==0.8.1